reader/requirements.txt
dacite

opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-flask
opentelemetry-exporter-otlp
opentelemetry-instrumentation-psycopg2